backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Caching==2.1.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
Flask-JWT-Extended==4.6.0

Flask-Migrate==4.0.4
Flask-Bcrypt==1.0.1
SQLAlchemy==2.0.19
celery==5.3.1
redis==4.6.0
requests==2.31.0
email-validator==2.2.0
Pillow==10.0.0
colorama==0.4.6
kombu==5.3.6